audio-to-text/requirements.txt
torch==2.0.0
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
google-cloud-storage==3.4.1
ffmpeg-python==0.2.0
openai-whisper==20250625